test: refactor product service tests with improved mock repository and error handling
